items

+++ PortalClickerApi/Database/Models/ClickerPlayer.cs
@@ -9,8 +9,8 @@ namespace PortalClickerApi.Database.Models
         public Guid UserId { get; set; }
         public ApplicationUser User { get; set; }
 
-        public ulong PortalCount { get; set; }
-        public ulong PortalsPerSecond { get; set; }
+        public double PortalCount { get; set; }
+        public double PortalsPerSecond { get; set; }
         public DateTime? LastTick { get; set; }
         public DateTime? LastClick { get; set; }
